Moved PartitionKey back to public scope
