+++ benchmark/src/main/java/com/azure/data/cosmos/benchmark/AsyncQueryBenchmark.java
@@ -26,7 +26,7 @@ package com.azure.data.cosmos.benchmark;
 import com.azure.data.cosmos.Document;
 import com.azure.data.cosmos.FeedOptions;
 import com.azure.data.cosmos.FeedResponse;
-import com.azure.data.cosmos.internal.PartitionKey;
+import com.azure.data.cosmos.PartitionKey;
 import reactor.core.publisher.BaseSubscriber;
 import reactor.core.publisher.Flux;
 import reactor.core.scheduler.Schedulers;
