+++ benchmark/src/main/java/com/azure/data/cosmos/benchmark/AsyncReadBenchmark.java
@@ -24,7 +24,7 @@
 package com.azure.data.cosmos.benchmark;
 
 import com.azure.data.cosmos.Document;
-import com.azure.data.cosmos.internal.PartitionKey;
+import com.azure.data.cosmos.PartitionKey;
 import com.azure.data.cosmos.internal.RequestOptions;
 import com.azure.data.cosmos.internal.ResourceResponse;
 import com.codahale.metrics.Timer;
