+++ commons/src/main/java/com/azure/data/cosmos/CosmosStoredProcedureRequestOptions.java
@@ -22,7 +22,6 @@
  */
 package com.azure.data.cosmos;
 
-import com.azure.data.cosmos.internal.PartitionKey;
 import com.azure.data.cosmos.internal.RequestOptions;
 
 /**
