+++ commons/src/test/java/com/azure/data/cosmos/PermissionTest.java
@@ -22,7 +22,6 @@
  */
 package com.azure.data.cosmos;
 
-import com.azure.data.cosmos.internal.PartitionKey;
 import com.azure.data.cosmos.internal.Permission;
 import org.testng.annotations.Test;
