+++ examples/src/test/java/com/azure/data/cosmos/rx/examples/TokenResolverTest.java
@@ -32,7 +32,7 @@ import com.azure.data.cosmos.Database;
 import com.azure.data.cosmos.Document;
 import com.azure.data.cosmos.DocumentClientTest;
 import com.azure.data.cosmos.DocumentCollection;
-import com.azure.data.cosmos.internal.PartitionKey;
+import com.azure.data.cosmos.PartitionKey;
 import com.azure.data.cosmos.PartitionKeyDefinition;
 import com.azure.data.cosmos.internal.Permission;
 import com.azure.data.cosmos.PermissionMode;
