+++ sdk/src/main/java/com/azure/data/cosmos/CosmosPermissionProperties.java
@@ -23,7 +23,6 @@
 package com.azure.data.cosmos;
 
 import com.azure.data.cosmos.internal.Constants;
-import com.azure.data.cosmos.internal.PartitionKey;
 import com.azure.data.cosmos.internal.Permission;
 import org.apache.commons.lang3.StringUtils;
