+++ sdk/src/test/java/com/azure/data/cosmos/directconnectivity/DCDocumentCrudTest.java
@@ -32,7 +32,7 @@ import com.azure.data.cosmos.Document;
 import com.azure.data.cosmos.DocumentCollection;
 import com.azure.data.cosmos.FeedOptions;
 import com.azure.data.cosmos.FeedResponse;
-import com.azure.data.cosmos.internal.PartitionKey;
+import com.azure.data.cosmos.PartitionKey;
 import com.azure.data.cosmos.internal.RequestOptions;
 import com.azure.data.cosmos.internal.ResourceResponse;
 import com.azure.data.cosmos.internal.StoredProcedure;
