+++ sdk/src/test/java/com/azure/data/cosmos/internal/ConsistencyTests2.java
@@ -33,6 +33,7 @@ import com.azure.data.cosmos.Document;
 import com.azure.data.cosmos.DocumentCollection;
 import com.azure.data.cosmos.FeedOptions;
 import com.azure.data.cosmos.FeedResponse;
+import com.azure.data.cosmos.PartitionKey;
 import com.azure.data.cosmos.directconnectivity.WFConstants;
 import com.azure.data.cosmos.internal.routing.PartitionKeyInternal;
 import com.azure.data.cosmos.rx.FailureValidator;
