+++ sdk/src/test/java/com/azure/data/cosmos/internal/ConsistencyTestsBase.java
@@ -34,6 +34,7 @@ import com.azure.data.cosmos.ConsistencyLevel;
 import com.azure.data.cosmos.Database;
 import com.azure.data.cosmos.Document;
 import com.azure.data.cosmos.DocumentCollection;
+import com.azure.data.cosmos.PartitionKey;
 import com.azure.data.cosmos.PartitionKeyDefinition;
 import com.azure.data.cosmos.PartitionKind;
 import com.azure.data.cosmos.Resource;
