+++ sdk/src/test/java/com/azure/data/cosmos/internal/DocumentQuerySpyWireContentTest.java
@@ -29,6 +29,7 @@ import com.azure.data.cosmos.Document;
 import com.azure.data.cosmos.DocumentCollection;
 import com.azure.data.cosmos.FeedOptions;
 import com.azure.data.cosmos.FeedResponse;
+import com.azure.data.cosmos.PartitionKey;
 import com.azure.data.cosmos.SpyClientBuilder;
 import com.azure.data.cosmos.internal.http.HttpRequest;
 import com.azure.data.cosmos.rx.Utils;
