+++ sdk/src/test/java/com/azure/data/cosmos/rx/OrderbyDocumentQueryTest.java
@@ -32,7 +32,7 @@ import com.azure.data.cosmos.CosmosItemProperties;
 import com.azure.data.cosmos.CosmosItemRequestOptions;
 import com.azure.data.cosmos.FeedOptions;
 import com.azure.data.cosmos.FeedResponse;
-import com.azure.data.cosmos.internal.PartitionKey;
+import com.azure.data.cosmos.PartitionKey;
 import com.azure.data.cosmos.Resource;
 import com.azure.data.cosmos.RetryAnalyzer;
 import com.azure.data.cosmos.internal.Utils.ValueHolder;
