+++ sdk/src/test/java/com/azure/data/cosmos/rx/ResourceTokenTest.java
@@ -29,7 +29,7 @@ import com.azure.data.cosmos.Database;
 import com.azure.data.cosmos.DatabaseForTest;
 import com.azure.data.cosmos.Document;
 import com.azure.data.cosmos.DocumentCollection;
-import com.azure.data.cosmos.internal.PartitionKey;
+import com.azure.data.cosmos.PartitionKey;
 import com.azure.data.cosmos.PartitionKeyDefinition;
 import com.azure.data.cosmos.internal.Permission;
 import com.azure.data.cosmos.PermissionMode;
