+++ sdk/src/test/java/com/azure/data/cosmos/rx/TestSuiteBase.java
@@ -54,7 +54,7 @@ import com.azure.data.cosmos.FeedResponse;
 import com.azure.data.cosmos.IncludedPath;
 import com.azure.data.cosmos.Index;
 import com.azure.data.cosmos.IndexingPolicy;
-import com.azure.data.cosmos.internal.PartitionKey;
+import com.azure.data.cosmos.PartitionKey;
 import com.azure.data.cosmos.PartitionKeyDefinition;
 import com.azure.data.cosmos.Resource;
 import com.azure.data.cosmos.RetryOptions;
