+++ sdk/src/test/java/com/azure/data/cosmos/rx/TokenResolverTest.java
@@ -35,7 +35,7 @@ import com.azure.data.cosmos.Document;
 import com.azure.data.cosmos.DocumentCollection;
 import com.azure.data.cosmos.FeedOptions;
 import com.azure.data.cosmos.FeedResponse;
-import com.azure.data.cosmos.internal.PartitionKey;
+import com.azure.data.cosmos.PartitionKey;
 import com.azure.data.cosmos.internal.Permission;
 import com.azure.data.cosmos.PermissionMode;
 import com.azure.data.cosmos.internal.RequestOptions;
